# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
        "/api/v1/profiles/update": {
            "put": {
                "tags": [
                    "UPDATE",
                    "Profile"
                ],
                "operationId": "updateCategory",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseProfileDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/categories/update": {
            "put": {
                "tags": [
                    "Category",
                    "UPDATE"
                ],
                "operationId": "updateCategory_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseCategoryDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/profiles/save": {
            "post": {
                "tags": [
                    "SAVE",
                    "Profile"
                ],
                "operationId": "saveProfile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProfileDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseProfileDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/profiles/all": {
            "get": {
                "tags": [
                    "FIND",
                    "Profile"
                ],
                "operationId": "getAllProfilesForUser",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ResponseProfileDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/profiles/delete/{id}": {
            "delete": {
                "tags": [
                    "DELETE",
                    "Profile"
                ],
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/categories/delete/{id}": {
            "delete": {
                "tags": [
                    "DELETE",
                    "Category"
                ],
                "operationId": "deleteCategory_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/accounts/delete/{id}": {
            "delete": {
                "tags": [
                    "Account",
                    "DELETE"
                ],
                "operationId": "deleteCategory_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateProfileDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "profileName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "profilePicture": {
                        "type": "string"
                    },
                    "profilePictureFile": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "required": [
                    "id"
                ]
            },
            "ResponseProfileDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "profileName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "profilePicture": {
                        "type": "string"
                    },
                    "accountIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "categoryIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    },
                    "expenseIds": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "uniqueItems": true
                    }
                }
            },
...
            "UpdateCategoryDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "id"
                ]
            },
            "ResponseCategoryDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "profileId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "colorCode",
                    "id",
                    "name",
                    "profileId"
                ]
            },
...
            "CreateProfileDTO": {
                "type": "object",
                "properties": {
                    "profileName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "colorCode": {
                        "type": "string"
                    },
                    "profilePicture": {
                        "type": "string"
                    },
                    "profilePictureFile": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "required": [
                    "colorCode",
                    "profileName"
                ]
            },
...
        }
    }
}